requirements.txt
# === Core Flask & Extensions ===
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
flask-jwt-extended>=4.6.0
gunicorn>=21.2.0
Werkzeug>=3.0.0

# === Database ===
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# === Task Queue (Celery) ===
celery>=5.3.4
redis>=5.0.1

# === Azure Services ===
azure-ai-formrecognizer>=3.3.2
azure-ai-documentintelligence>=1.0.0b1
azure-core>=1.29.5

# === AI & LLM ===
openai>=1.3.7
anthropic>=0.8.0

# === Document Processing ===
pymupdf>=1.23.7
pillow>=10.1.0
opencv-python-headless>=4.8.1.78

# === Excel Processing ===
openpyxl>=3.1.2
pandas>=2.1.3

# === Utilities ===
python-dateutil>=2.8.2
python-dotenv>=1.0.0
